+++ src/pages/exploiter/generator/IndexPage.vue
@@ -3,7 +3,7 @@ import { ref, reactive, onMounted } from 'vue'
 import type { FormInstance, FormRules } from 'element-plus'
 import draggable from 'vuedraggable'
 import DialogView from 'components/DialogView.vue'
-import CodeView from 'components/CodeView.vue'
+import CodeRender from 'components/CodeRender.vue'
 import { retrieveSchemas, retrieveSchemaFields, retrieveSchemaCodes, fetchSchema } from 'src/api/schemas'
 import type { Pagination, Schema, Field, Code } from 'src/models'
 
@@ -395,7 +395,7 @@ function handleCheckedChange(value: string[]) {
     :show-full-screen="true">
     <ElTabs stretch>
       <ElTabPane v-for="code in codes" :key="code.name" :label="code.name" class="w-full">
-        <CodeView :content="code.content" />
+        <CodeRender :content="code.content" />
       </ElTabPane>
     </ElTabs>
   </DialogView>
