+++ src/pages/exploiter/templates/IndexPage.vue
@@ -4,7 +4,7 @@ import type { FormInstance, FormRules } from 'element-plus'
 import { dayjs } from 'element-plus'
 import draggable from 'vuedraggable'
 import DialogView from 'components/DialogView.vue'
-import CodeView from 'components/CodeView.vue'
+import CodeRender from 'components/CodeRender.vue'
 import { retrieveTemplates, fetchTemplate } from 'src/api/templates'
 import type { Pagination, Template } from 'src/models'
 
@@ -308,6 +308,6 @@ function handleCheckedChange(value: string[]) {
   <!-- preview -->
   <DialogView v-model="previewDialogVisible" :title="$t('preview')" width="50%" :max-height="600"
     :show-full-screen="true">
-    <CodeView :content="form.content" />
+    <CodeRender :content="form.content" />
   </DialogView>
 </template>
